xgboost/requirements.txt
boto3~=1.17.58
pillow~=8.1.2
torch~=1.8.0
torchvision~=0.9.0
argparse~=1.4.0
pandas~=1.1.5
thop~=0.0.31-2005241907
tqdm~=4.59.0
numpy~=1.19.5
opencv-python~=4.5.1.48
matplotlib~=3.3.4
s3fs~=2021.4.0
scipy~=1.5.4
thop